pythonApp/requirements.txt
# $ pip install -r requirements.txt
opencv-python==4.1.2.30
pytesseract==0.3.1
watchdog==0.9.0
numpy==1.18.1
Pillow==7.0.0
